+++ src/shared/constants.js
@@ -2,7 +2,9 @@ export const DEFAULT_CALLBACKS = {
   onStart: () => {},
   onPreviousStep: (currentStep) => {},
   onNextStep: (currentStep) => {},
-  onStop: () => {}
+  onStop: () => {},
+  onSkip: () => {},
+  onFinish: () => {}
 }
 
 export const STICKY = {
@@ -10,24 +12,46 @@ export const STICKY = {
 }
 
 export const DEFAULT_OPTIONS = {
-  useKeyboardNavigation: true,
-  startTimeout: 0,
+  highlight: false,
   labels: {
     buttonSkip: 'Skip tour',
     buttonPrevious: 'Previous',
     buttonNext: 'Next',
     buttonStop: 'Finish'
-  }
+  },
+  enabledButtons: {
+    buttonSkip: true,
+    buttonPrevious: true,
+    buttonNext: true,
+    buttonStop: true
+  },
+  startTimeout: 0,
+  useKeyboardNavigation: true,
+  debug: false
+}
+
+export const HIGHLIGHT = {
+  CLASSES: {
+    ACTIVE: 'v-tour--active',
+    TARGET_HIGHLIGHTED: 'v-tour__target--highlighted',
+    TARGET_RELATIVE: 'v-tour__target--relative'
+  },
+  TRANSITION: 'box-shadow 0s ease-in-out 0s',
+  useKeyboardNavigation: true,
+  startTimeout: 0,
+  stopOnTargetNotFound: true
 }
 
 export const DEFAULT_STEP_OPTIONS = {
-  placement: 'bottom',
   enableScrolling: true,
+  highlight: DEFAULT_OPTIONS.highlight, // By default use the global tour setting
+  enabledButtons: DEFAULT_OPTIONS.enabledButtons,
   modifiers: {
     arrow: {
       element: '.v-step__arrow'
     }
-  }
+  },
+  placement: 'bottom'
 }
 
 export const KEYS = {