+++ tests/unit/VStep.spec.js
@@ -1,6 +1,7 @@
-import { expect } from 'chai'
-import { shallow } from '@vue/test-utils'
+import { shallowMount } from '@vue/test-utils'
 import VStep from '@/components/VStep.vue'
+
+const labels = {}
 
 describe('VStep.vue', () => {
   it('renders props.step.content', () => {
@@ -9,13 +10,14 @@ describe('VStep.vue', () => {
       content: 'This is a demo step!'
     }
 
-    const wrapper = shallow(VStep, {
+    const wrapper = shallowMount(VStep, {
       propsData: {
         step,
-        stop: () => {}
+        stop: () => {},
+        labels
       }
     })
 
-    expect(wrapper.text()).to.include(step.content)
+    expect(wrapper.text()).toContain(step.content)
   })
 })